creating repo

+++ cogs/moderation.py
@@ -8,12 +8,15 @@ class moderation(commands.Cog):
         self.bot = bot
 
     @commands.command()
-    @commands.has_role(roles['Admin'], roles['Moderator'])
-    async def clear(self, ctx, amount):
-        await ctx.channel.purge(limit=amount)
-        msg = await ctx.send("*Сообщения удалены...*")
-        await asyncio.sleep(2)
-        await msg.delete()
+    @commands.has_any_role(roles['Admin'], roles['Moderator'])
+    async def clear(self, ctx, amount: int):
+        if not amount:
+            await ctx.send("Вы не указали кол-во сообщений")
+        else:
+            await ctx.channel.purge(limit=amount)
+            msg = await ctx.send("*Сообщения удалены...*")
+            await asyncio.sleep(2)
+            await msg.delete()
 
 
 def setup(bot):
